test: extract site explorer power shelf tests
Move power shelf integration coverage out of api-core site explorer
tests into the site-explorer crate, and introduce a shared test
harness foundation for cross-crate test setup.
