+++ crates/api-core/src/tests/common/api_fixtures/mod.rs
@@ -128,18 +128,17 @@ use crate::test_support::builder::TestApiBuilder;
 use crate::test_support::default_config;
 use crate::test_support::ib_fabric::ib_fabric_test_manager;
 pub use crate::test_support::network::{FIXTURE_DHCP_RELAY_ADDRESS, TEST_SITE_PREFIXES};
-pub use crate::test_support::network_segment;
 use crate::test_support::network_segment::{
     FIXTURE_TENANT_NETWORK_SEGMENT_GATEWAYS, create_admin_network_segment,
     create_static_assignments_segment, create_tenant_network_segment,
     create_underlay_network_segment,
 };
+pub use crate::test_support::{endpoint_explorer, network_segment};
 use crate::tests::common::api_fixtures::endpoint_explorer::MockEndpointExplorer;
 use crate::tests::common::api_fixtures::managed_host::ManagedHostConfig;
 use crate::tests::common::rpc_builder::VpcCreationRequest;
 
 pub mod dpu;
-pub mod endpoint_explorer;
 pub mod host;
 pub mod ib_partition;
 pub mod instance;
