+++ crates/site-explorer/Cargo.toml
@@ -50,8 +50,14 @@ thiserror = { workspace = true }
 version-compare = { workspace = true }
 
 [dev-dependencies]
+bmc-vendor = { path = "../bmc-vendor" }
+carbide-api-core = { path = "../api-core", features = ["test-support"] }
+carbide-test-harness = { path = "../test-harness" }
 carbide-redfish = { path = "../redfish", features = ["test-support"] }
+carbide-rpc = { path = "../rpc", features = ["test-support"] }
+
 tokio = { workspace = true, features = ["macros", "rt"] }
+tonic = { workspace = true }
 
 [lints]
 workspace = true
